fix: resolved importing and styling comments
