+++ client/package.json
@@ -22,7 +22,8 @@
         "framer-motion": "^11.2.13",
         "lint-staged": "^14.0.1",
         "react": "^18.2.0",
-        "react-dom": "^18.2.0"
+        "react-dom": "^18.2.0",
+        "react-router-dom": "^7.8.0"
     },
     "lint-staged": {
         "*.{js,jsx}": "yarn run eslint"
